requirements.txt
sqlalchemy~=2.0.41
psycopg[binary]
passlib~=1.7.4
fastapi~=0.116.1
pydantic~=2.11.7
python-dotenv~=1.1.1
uvicorn[standard]~=0.27.1
python-jose[cryptography]~=3.5.0
